wip: extract observability pricing
